+++ congress/CHANGELOG.md
@@ -28,6 +28,10 @@ Change log entries are to be added to the Unreleased section. Example entry:
 
 ## v1.2.0 (2023-12-28)
 
+### Features
+
+- `all_hook_permissions` query to return all hook permissions.
+
 ### Improvements
 
 - Remove temporal `add_tc_dismiss_perm` and gas adjustments in the `execute` method related to the wrong Gas settings in the TC proposals (and missing checks mentioned in the v1.1.2 release).